KAN-269 fix: 주문 데이출력 오류 수정
