+++ ubun2_be/src/main/java/kr/or/kosa/ubun2_be/domain/order/service/OrderService.java
@@ -11,8 +11,6 @@ public interface OrderService {
 
     OrderDetailResponse getOrderByCustomerIdAndOrderId(Long orderId, Long customerId);
 
-    SubscriptionOrderDetailResponse getSubscriptionOrderByCustomerIdAndOrderId(Long orderId, Long customerId);
-
     Page<UnifiedOrderResponse> getPendingOrders(Long customerId, SearchRequest searchRequest, Pageable pageable);
 
     void updateOrderStatus(Long customerId, List<OrderApproveRequest> orderApproveRequests);
