+++ ubun2_be/src/main/java/kr/or/kosa/ubun2_be/domain/paymentmethod/repository/CardPaymentRepository.java
@@ -7,4 +7,6 @@ import java.util.Optional;
 
 public interface CardPaymentRepository extends JpaRepository<CardPayment, Long> {
     Optional<CardPayment> findByIsDeletedFalseAndPaymentMethodId(Long paymentMethodId);
+
+    Optional<CardPayment> findByPaymentMethodId(Long paymentMethodId);
 }
